# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoWVP Owl 的 HTTP API 文档，包含设备、通道、GB28181、PTZ、录像与系统接口。\n\n鉴权方式：\n大多数管理接口需要在请求头中携带 `Authorization: Bearer \u003ctoken\u003e`。\n可先调用 `/login/key` 获取 RSA 公钥，再调用 `/login` 完成登录。\n\n返回约定：\n1. 成功时通常返回业务对象，或 `{ \"msg\": \"ok\" }`\n2. 失败时通常返回 `{ \"msg\": \"错误原因\" }`\n3. GB28181 查询类接口通常会同时返回结构化结果与原始 XML",
        "title": "GoWVP Owl API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/ai/events": {
            "post": {
                "description": "AI 服务上报检测结果。平台会按 `detections` 中每个对象分别落事件记录，并保存快照图片。\n请求示例：`{ \"camera_id\":\"GB_34020000001320000001\", \"timestamp\":1710931200000, \"detections\":[{\"label\":\"person\",\"confidence\":0.92,\"box\":{\"x_min\":120,\"y_min\":80,\"x_max\":360,\"y_max\":500},\"area\":35800}], \"snapshot_width\":1920, \"snapshot_height\":1080 }`\n成功响应示例：`{ \"code\": 0, \"msg\": \"success\" }`",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIWebhook"
                ],
                "summary": "AI 检测事件回调",
                "parameters": [
                    {
                        "description": "AI 检测结果",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AIDetectionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.AIWebhookOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "查询事件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "did",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "通道ID",
                        "name": "cid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签",
                        "name": "label",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始毫秒时间戳",
                        "name": "start_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束毫秒时间戳",
                        "name": "end_ms",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/image/{path}": {
            "get": {
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "获取事件快照图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图片相对路径",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JPEG 图片",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "获取事件详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "事件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "修改事件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "事件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "事件更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.EditEventInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "删除事件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "事件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AIBoundingBox": {
            "type": "object",
            "properties": {
                "x_max": {
                    "type": "integer",
                    "example": 360
                },
                "x_min": {
                    "type": "integer",
                    "example": 120
                },
                "y_max": {
                    "type": "integer",
                    "example": 500
                },
                "y_min": {
                    "type": "integer",
                    "example": 80
                }
            }
        },
        "api.AIDetection": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "边界框像素面积",
                    "type": "integer",
                    "example": 35800
                },
                "box": {
                    "description": "像素坐标边界框",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.AIBoundingBox"
                        }
                    ]
                },
                "confidence": {
                    "description": "置信度（0.0-1.0）",
                    "type": "number",
                    "example": 0.92
                },
                "label": {
                    "description": "物体类别",
                    "type": "string",
                    "example": "person"
                },
                "norm_box": {
                    "description": "归一化边界框",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.AINormBox"
                        }
                    ]
                }
            }
        },
        "api.AIDetectionInput": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "description": "摄像头 ID",
                    "type": "string",
                    "example": "GB_34020000001320000001"
                },
                "detections": {
                    "description": "检测结果列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AIDetection"
                    }
                },
                "snapshot": {
                    "description": "Base64 编码的快照（JPEG）",
                    "type": "string"
                },
                "snapshot_height": {
                    "description": "快照高度",
                    "type": "integer",
                    "example": 1080
                },
                "snapshot_width": {
                    "description": "快照宽度",
                    "type": "integer",
                    "example": 1920
                },
                "timestamp": {
                    "description": "Unix 时间戳 (毫秒)",
                    "type": "string"
                }
            }
        },
...
        "api.AINormBox": {
            "type": "object",
            "properties": {
                "h": {
                    "description": "高度",
                    "type": "number",
                    "example": 0.32
                },
                "w": {
                    "description": "宽度",
                    "type": "number",
                    "example": 0.18
                },
                "x": {
                    "description": "中心点 X 坐标",
                    "type": "number",
                    "example": 0.42
                },
                "y": {
                    "description": "中心点 Y 坐标",
                    "type": "number",
                    "example": 0.51
                }
            }
        },
...
        "api.AIWebhookOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误代码，0 表示成功",
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "description": "消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
        "api.SwaggerEventsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "事件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event.Event"
                    }
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
...
        "event.EditEventInput": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "description": "事件结束时间 (毫秒时间戳)",
                    "type": "string"
                }
            }
        },
        "event.Event": {
            "type": "object",
            "properties": {
                "cid": {
                    "description": "通道 ID (channel 表的 id)",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "did": {
                    "description": "设备 ID (device 表的 id)",
                    "type": "string"
                },
                "ended_at": {
                    "description": "事件结束时间",
                    "type": "string"
                },
                "id": {
                    "description": "事件 ID",
                    "type": "integer"
                },
                "image_path": {
                    "description": "图片相对路径 (cid/年月日时分秒_随机6位.jpg)",
                    "type": "string"
                },
                "label": {
                    "description": "检测标签 (person, car 等)",
                    "type": "string"
                },
                "model": {
                    "description": "分析模型名称",
                    "type": "string"
                },
                "score": {
                    "description": "置信度 (0.0-1.0)",
                    "type": "number"
                },
                "started_at": {
                    "description": "事件开始时间",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "zones": {
                    "description": "检测区域，JSON 格式存储边界框信息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}